Change all visible occurrences of hateos->hateoas
