+++ http-client/src/main/java/io/micronaut/http/client/HttpClient.java
@@ -24,7 +24,7 @@ import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.HttpStatus;
 import io.micronaut.http.client.exceptions.HttpClientResponseException;
-import io.micronaut.http.hateos.JsonError;
+import io.micronaut.http.hateoas.JsonError;
 import org.reactivestreams.Publisher;
 
 import java.io.Closeable;
