+++ http-client/src/test/groovy/io/micronaut/http/client/aop/HateosCrudSpec.groovy
@@ -19,8 +19,8 @@ import io.micronaut.context.ApplicationContext
 import io.micronaut.http.MediaType
 import io.micronaut.http.annotation.*
 import io.micronaut.http.client.annotation.Client
-import io.micronaut.http.hateos.AbstractResource
-import io.micronaut.http.hateos.Link
+import io.micronaut.http.hateoas.AbstractResource
+import io.micronaut.http.hateoas.Link
 import io.micronaut.runtime.server.EmbeddedServer
 import spock.lang.AutoCleanup
 import spock.lang.Shared
@@ -33,7 +33,7 @@ import java.util.concurrent.atomic.AtomicLong
  * @since 1.0
  */
 
-class HateosCrudSpec extends Specification {
+class hateoasCrudSpec extends Specification {
 
     @Shared
     @AutoCleanup
@@ -88,11 +88,11 @@ class HateosCrudSpec extends Specification {
         client.get(book.id) == null
     }
 
-    @Client('/hateos/books')
+    @Client('/hateoas/books')
     static interface BookClient extends BookApi {
     }
 
-    @Controller("/hateos/books")
+    @Controller("/hateoas/books")
     static class BookController implements BookApi {
 
         Map<Long, Book> books = new LinkedHashMap<>()
