+++ http-server-netty/src/test/groovy/io/micronaut/http/server/netty/binding/JsonBodyBindingSpec.groovy
@@ -26,8 +26,8 @@ import io.micronaut.http.HttpStatus
 import io.micronaut.http.annotation.Body
 import io.micronaut.http.annotation.Error
 import io.micronaut.http.client.exceptions.HttpClientResponseException
-import io.micronaut.http.hateos.Link
-import io.micronaut.http.hateos.JsonError
+import io.micronaut.http.hateoas.Link
+import io.micronaut.http.hateoas.JsonError
 import io.micronaut.http.server.netty.AbstractMicronautSpec
 import io.micronaut.http.annotation.Controller
 import io.micronaut.http.annotation.Post
