+++ http-server/src/main/java/io/micronaut/http/server/exceptions/HttpStatusHandler.java
@@ -21,8 +21,8 @@ import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.annotation.Produces;
 import io.micronaut.http.exceptions.HttpStatusException;
-import io.micronaut.http.hateos.Link;
-import io.micronaut.http.hateos.JsonError;
+import io.micronaut.http.hateoas.Link;
+import io.micronaut.http.hateoas.JsonError;
 
 import javax.inject.Singleton;
