+++ http/src/main/java/io/micronaut/http/hateoas/package-info.java
@@ -14,32 +14,10 @@
  * limitations under the License.
  */
 
-package io.micronaut.http.hateos;
-
-import io.micronaut.core.annotation.Internal;
-import io.micronaut.http.MediaType;
-import io.micronaut.http.annotation.Produces;
-
 /**
- * A type that can be used to represent JSON errors that returns a content type of Vnd.Error.
+ * hateoas.
  *
- * @author graemerocher
+ * @author Graeme Rocher
  * @since 1.0
  */
-@Produces(MediaType.APPLICATION_VND_ERROR)
-public class VndError extends JsonError {
-
-    /**
-     * @param message The message
-     */
-    public VndError(String message) {
-        super(message);
-    }
-
-    /**
-     * Used by Jackson.
-     */
-    @Internal
-    VndError() {
-    }
-}
+package io.micronaut.http.hateoas;
