+++ src/main/docs/guide/httpServer/errorHandling.adoc
@@ -19,7 +19,7 @@ include::{testsuite}/server/json/PersonController.java[tags=localError, indent=0
 ----
 
 <1> A method that explicitly handles `JsonParseException` is declared
-<2> An instance of api:http.hateos.JsonError[] is returned.
+<2> An instance of api:http.hateoas.JsonError[] is returned.
 <3> A custom response is returned to handle the error
 
 .Local status handler
@@ -30,7 +30,7 @@ include::{testsuite}/server/json/PersonController.java[tags=statusError, indent=
 ----
 
 <1> The api:http.annotation.Error[] declares which api:http.HttpStatus[] error code to handle (in this case 404)
-<2> A api:http.hateos.JsonError[] instance is returned for all 404 responses
+<2> A api:http.hateoas.JsonError[] instance is returned for all 404 responses
 <3> An api:http.HttpStatus#NOT_FOUND[] response is returned
 
 == Global Error Handling
@@ -43,7 +43,7 @@ include::{testsuite}/server/json/PersonController.java[tags=globalError, indent=
 ----
 
 <1> The api:http.annotation.Error[] is used to declare the method a global error handler
-<2> A api:http.hateos.JsonError[] instance is returned for all errors
+<2> A api:http.hateoas.JsonError[] instance is returned for all errors
 <3> An api:http.HttpStatus#INTERNAL_SERVER_ERROR[] response is returned
 
 .Global status handler
@@ -54,7 +54,7 @@ include::{testsuite}/server/json/PersonController.java[tags=statusError, indent=
 ----
 
 <1> The api:http.annotation.Error[] declares which api:http.HttpStatus[] error code to handle (in this case 404)
-<2> A api:http.hateos.JsonError[] instance is returned for all 404 responses
+<2> A api:http.hateoas.JsonError[] instance is returned for all 404 responses
 <3> An api:http.HttpStatus#NOT_FOUND[] response is returned
 
 IMPORTANT: A few things to note about the api:http.annotation.Error[] annotation. Two identical `@Error` annotations that are
