+++ validation/src/main/java/io/micronaut/validation/exceptions/ValidationExceptionHandler.java
@@ -19,8 +19,8 @@ package io.micronaut.validation.exceptions;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
-import io.micronaut.http.hateos.JsonError;
-import io.micronaut.http.hateos.Link;
+import io.micronaut.http.hateoas.JsonError;
+import io.micronaut.http.hateoas.Link;
 import io.micronaut.http.server.exceptions.ExceptionHandler;
 import org.grails.datastore.mapping.validation.ValidationException;
 import org.springframework.validation.Errors;
